# ergosense-app/server/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ErgoSense API",
    "version": "1.0.0",
    "description": "API REST multi-tenant para SST, ergonomia, GRO, PGR, AET, psicossocial, eSocial e compliance."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Gateway relativo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/denuncias/dashboard": {
      "get": {
        "tags": [
          "denuncias"
        ],
        "summary": "GET /api/denuncias/dashboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/api/denuncias": {
      "get": {
        "tags": [
          "denuncias"
        ],
        "summary": "GET /api/denuncias",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      },
      "post": {
        "tags": [
          "denuncias"
        ],
        "summary": "POST /api/denuncias",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/denuncias/:id": {
      "get": {
        "tags": [
          "denuncias"
        ],
        "summary": "GET /api/denuncias/:id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/api/denuncias/public": {
      "post": {
        "tags": [
          "denuncias"
        ],
        "summary": "POST /api/denuncias/public",
        "security": [],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/denuncias/public/status": {
      "get": {
        "tags": [
          "denuncias"
        ],
        "summary": "GET /api/denuncias/public/status",
        "security": [],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/api/denuncias/:id/status": {
      "patch": {
        "tags": [
          "denuncias"
        ],
        "summary": "PATCH /api/denuncias/:id/status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/denuncias/:id/tratativas": {
      "post": {
        "tags": [
          "denuncias"
        ],
        "summary": "POST /api/denuncias/:id/tratativas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/denuncias/:id/tratativas/:tid": {
      "patch": {
        "tags": [
          "denuncias"
        ],
        "summary": "PATCH /api/denuncias/:id/tratativas/:tid",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/denuncias/:id/evidencias": {
      "post": {
        "tags": [
          "denuncias"
        ],
        "summary": "POST /api/denuncias/:id/evidencias",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/denuncias/:id/integrar": {
      "post": {
        "tags": [
          "denuncias"
        ],
        "summary": "POST /api/denuncias/:id/integrar",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/denuncias/:id/conclusao": {
      "patch": {
        "tags": [
          "denuncias"
        ],
        "summary": "PATCH /api/denuncias/:id/conclusao",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  }
}